Fixed bug in alignment_functions; allow user to specify fastq file to be aligned (but file must already be within the container)

+++ align_genome/scripts/cleanup_functions.py
@@ -2,19 +2,29 @@
 
 import os
 
-def clear_dir(dirpath):
+def clear_dir(dirpath, keep_flag):
 
     """
     Clear all files in a directory.
+    If keep_flag is True, keep the final files (sorted bam and index).
 
     :param dirpath: Path to directory of interest.
     """
 
     all_files = os.listdir(dirpath)
 
-    for file in all_files:
-        full_path = f'{dirpath}/{file}'
-        os.remove(full_path)
+    if not keep_flag:
+        for file in all_files:
+            full_path = f'{dirpath}/{file}'
+            os.remove(full_path)
 
-    assert len(os.listdir(dirpath)) == 0, "Problem with clean-up: \
+        assert len(os.listdir(dirpath)) == 0, "Problem with clean-up: \
                                           not all files removed."
+    else:
+        for file in all_files:
+            full_path = f'{dirpath}/{file}'
+            if not "sorted" in full_path:
+                os.remove(full_path)
+
+        assert len(os.listdir(dirpath)) == 2, "Problem with clean-up: \
+                                                final number of files different from desired."
